Refactor Scala 2 expression compiler
